Refactor program
rename Program1 to BlazorServerStarter, as we already got a Program but I needed a better name for that
fix minor codestyle issues
