+++ src/AasxServerBlazor/App.razor
@@ -1,4 +1,4 @@
-﻿<Router AppAssembly="@typeof(Program1).Assembly">
+﻿<Router AppAssembly="@typeof(BlazorServerStarter).Assembly">
     <Found Context="routeData">
         <RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
     </Found>
